added an opportunity to add brands preferences to showroom while registrating
